// apps/server/src/resources/Message.ts
export enum Message {
    AppNamePrefix = 'MyShipServer',
    NO_LICENCE_INFO = 'Es sind keine Lizenzinformationen in der Konfig hinterlegt!',
    UNKNOWN_ERROR = 'Es ist ein unerwarteter Fehler aufgetreten, dies sollte allerdings keinen Einfluss auf die Funktionalität haben.',
    INVALID_LICENCE = 'Der eingetragene Lizenzschlüssel ist nicht gültig!',
    CLOUD_SYSTEM_TIMEOUT = 'Eine Verbindung zum Cloud-System ist nicht möglich, es kann zu funktionalen Einschränkungen kommen.',
    CLOUD_SYSTEM_CONNECTED = 'Die Lizenz ist gültig und ausgestellt auf: ',
    SERVER_LISTEN_AT = 'Der Server lauscht auf Port:',
    SERVER_STARTING = 'Der Server wird nun gestartet!',
    DATABASE_ERROR = 'Es konnte keine Datenbankverbindung hergestellt werden oder die erforderliche Datenbankstruktur ist fehlerhaft!',
    SCHEDULE_LOGGING = 'Der Server wird nun regelmäßig die BlackBox überprüfen.',
    WS_CLIENT_CONNECTED = 'Ein Client hat eine WebSocket-Verbindung mit dem Server aufgebaut. #',
    SEND_BRIDGE_REQUEST = 'Es wurde eine Anfrage zu Bridge geschickt.',
    LOG_TRACK_MANUAL = 'Ein Track des Logbuchs wurde manuell aktualisiert. #',
    LOG_TRACK = 'Ein Track des Logbuchs wurde aktualisiert. #',
    LOG_TRACK_ERROR = 'Ein Track konnte nicht aktualisiert werden. #',
    NO_SIGNALK_CONNECTION = 'Es konnte keine Verbindung zum SignalK-Server aufgebaut werden!',
    LOGGING_ACTIVE_TRACKS = 'Aktive Tracks werden gerade geloggt.',
    UPDATE_APT_ERROR = "Linux Pakete konnten nicht aktualisiert werden!",
    UP_TO_DATE = "Die Software befindet sich auf dem aktuellsten Stand!",
    PULL_NEW_VERSION = "Neue Version wird aus dem Respository geladen...",
    FAILED_UPDATE_NPM_PACKAGES = "NPM-Pakete konnte erfolgreich geladen werden!",
    SUCCESS_UPDATE_NPM_PACKAGES = "NPM-Pakete wurden erfolgreich geladen!",
    FAILED_UPDATE_LINUX_PACKAGES = "Linux-Paketlichen konnte nicht aktualisiert werden!",
    SUCCESS_UPDATE_LINUX_PACKAGES = "Linux-Paketlisten wurden erfolgreich aktualisiert!",
    RESTART_IN_PROGRESS = "Der Server wird in kürze neugestartet!",
    DATABASE_CONNECTION_FAILED = "Es konnte keine Datenbankverbindung aufgebaut werden, bitte Konfiguration überprüfen!",
    MQTT_STARED = "Der MQTT-Broker wurde gestartet!",
    CLOUD_SIGNALK_KEYS_SYNC_FAILED = "Die SignalK-Keys konnten nicht mit der Cloud synchronisiert werden!",
    CLOUD_SIGNALK_KEYS_LOAD_FAILED = "Die SignalK-Keys konnten nicht aus der Cloud geladen werden!",
    SIGNALK_DATA_LOAD_FAILED = "Die SignalK-Daten konnten durch den Pfad nicht geladen werden!",
    FAILED_TO_BACKUP_TRACK = "Der Track konnte nicht gesichert werden!",
}